fixed set_primary_block_bounds(), lowest_primary
defined ILLEGAL16 as '(short) x0ffff'
  could not find out why the '(short)' is needed...

modified:   pulses/infos/TODO.txt
modified:   pulses/men_MuD.h
modified:   pulses/musicBox.h
modified:   pulses/pulses.ino

+++ pulses/pulses.ino
@@ -55,9 +55,9 @@ using namespace std;	// ESP8266 needs that
 /* **************************************************************** */
 // some #define's:
 
-#define ILLEGAL8	255
-#define ILLEGAL16	0xffff
-#define ILLEGAL32	0xffffffff
+#define ILLEGAL8	(uint8_t) 255
+#define ILLEGAL16	(short) 0xffff
+#define ILLEGAL32	(int) 0xffffffff
 
 /* **************************************************************** */
 // configuration sequence:
